🏗️ Move types to improve readability
