+++ src/index.ts
@@ -3,33 +3,8 @@ import {
   join, posix, relative, sep,
 } from 'path';
 import { writeFileSync, mkdirSync, readFileSync } from 'fs';
-import { createManifest, CreateManifestOptions } from './create';
-
-export interface VitePluginCustomElementsManifestOptions extends CreateManifestOptions {
-  /**
-   * Define where will be serve the manifest.
-   * This option only work in development mode.
-   * @default '/custom-elements.json'
-   */
-  endpoint?: string,
-  /**
-   * Define where will be build the final manifest.
-   * This option only work in production mode.
-   * @default 'custom-elements.json'
-   */
-  output?: string,
-  /**
-   * Add the custom-elements-manifest field to the package.json.
-   * @default false
-   */
-  packageJson?: boolean,
-  /**
-   * Register files which will be used to build the manifest.
-   * You can use pattern to find files.
-   * @default []
-   */
-  files?: string[],
-}
+import { createManifest } from './create';
+import { VitePluginCustomElementsManifestOptions } from './types';
 
 function VitePluginCustomElementsManifest({
   endpoint = '/custom-elements.json',
@@ -43,6 +18,7 @@ function VitePluginCustomElementsManifest({
 
   return {
     name: 'vite-plugin-custom-elements-manifest',
+
     configureServer(server) {
       server.middlewares.use(endpoint, async (req, res) => {
         const manifest = createManifest(files, createManifestOptions);
@@ -50,6 +26,7 @@ function VitePluginCustomElementsManifest({
         res.end(JSON.stringify(manifest));
       });
     },
+
     generateBundle(this, { dir }) {
       const path = join(dir, output);
       const manifest = createManifest(files, createManifestOptions);
@@ -67,6 +44,7 @@ function VitePluginCustomElementsManifest({
         writeFileSync(packageJsonPath, `${JSON.stringify(pkg, null, 2)}\n`);
       }
     },
+
     resolveId(id) {
       if (id !== virtualModuleId) {
         return undefined;
@@ -74,6 +52,7 @@ function VitePluginCustomElementsManifest({
 
       return resolvedVirtualModuleId;
     },
+
     load(id) {
       if (id !== resolvedVirtualModuleId) {
         return undefined;
@@ -82,6 +61,7 @@ function VitePluginCustomElementsManifest({
       const manifest = createManifest(files, createManifestOptions);
       return `export default ${JSON.stringify(manifest)}`;
     },
+
     async handleHotUpdate({ server }) {
       const mod = await server.moduleGraph.getModuleByUrl(resolvedVirtualModuleId);
